VM integration + reusable error component (#9) - Implemented new state management that connects the question and its metadata - Implemented a reusable composable that displays errors

+++ Software/ErasMapp/app/src/main/java/com/rampu/erasmapp/channels/ui/questions/QuestionListItem.kt
@@ -7,5 +7,5 @@ data class QuestionListItem(
     val authorLabel: String,
     val authorPhotoUrl: String?,
     val lastActivityAt: Long?,
-    val unreadCount: Int
+    val unreadCount: Long
 )
